tests(Testcafe_devextreme): split heavy data grid test into different fixtures
