+++ e2e/testcafe-devextreme/tests/dataGrid/stickyColumns/helpers/data.ts
