+++ e2e/testcafe-devextreme/tests/dataGrid/stickyColumns/helpers/testMatrixConfig.ts
@@ -1,0 +1,51 @@
+export const borderConfigs = [
+  {
+    showRowLines: true,
+    showColumnLines: true,
+    showBorders: true,
+  }, {
+    showRowLines: false,
+    showColumnLines: true,
+    showBorders: true,
+  }, {
+    showRowLines: false,
+    showColumnLines: false,
+    showBorders: true,
+  }, {
+    showRowLines: false,
+    showColumnLines: false,
+    showBorders: false,
+  }, {
+    showRowLines: true,
+    showColumnLines: false,
+    showBorders: true,
+  }, {
+    showRowLines: true,
+    showColumnLines: false,
+    showBorders: false,
+  }, {
+    showRowLines: true,
+    showColumnLines: true,
+    showBorders: false,
+  }, {
+    showRowLines: false,
+    showColumnLines: true,
+    showBorders: false,
+  },
+];
+
+export const rtlAndRowAltConfigs = [
+  {
+    rtlEnabled: true,
+    rowAlternationEnabled: true,
+  }, {
+    rtlEnabled: false,
+    rowAlternationEnabled: true,
+  }, {
+    rtlEnabled: false,
+    rowAlternationEnabled: false,
+  }, {
+    rtlEnabled: true,
+    rowAlternationEnabled: false,
+  },
+];
